+++ Src/BinaryReader.cpp
@@ -25,7 +25,8 @@ BinaryReader::BinaryReader(_In_z_ wchar_t const* fileName) noexcept(false) :
     HRESULT hr = ReadEntireFile(fileName, mOwnedData, &dataSize);
     if (FAILED(hr))
     {
-        DebugTrace("ERROR: BinaryReader failed (%08X) to load '%ls'\n", hr, fileName);
+        DebugTrace("ERROR: BinaryReader failed (%08X) to load '%ls'\n",
+            static_cast<unsigned int>(hr), fileName);
         throw std::exception("BinaryReader");
     }
